+++ test/integration/editors/subscription/goose/GooseControlSubscription.ts
@@ -2,22 +2,27 @@ import { html, fixture, expect } from '@open-wc/testing';
 
 import '../../../../mock-wizard.js';
 
-import SampledValues from '../../../../../src/editors/SampledValues.js';
+import GooseControlSubscriptionPlugin from '../../../../../src/editors/GooseControlSubscription.js';
 import { Editing } from '../../../../../src/Editing.js';
 import { Wizarding } from '../../../../../src/Wizarding.js';
 import { ListItem } from '@material/mwc-list/mwc-list-item.js';
 
-describe('Sampled Values Plugin', () => {
-  customElements.define('smv-plugin', Wizarding(Editing(SampledValues)));
-  let element: SampledValues;
+describe('Subscription Plugin', () => {
+  customElements.define(
+    'subscription-plugin',
+    Wizarding(Editing(GooseControlSubscriptionPlugin))
+  );
+  let element: GooseControlSubscriptionPlugin;
   let doc: XMLDocument;
 
   beforeEach(async () => {
-    doc = await fetch('/test/testfiles/valid2007B4ForSampledValues.scd')
+    doc = await fetch('/test/testfiles/valid2007B4ForSubscription.scd')
       .then(response => response.text())
       .then(str => new DOMParser().parseFromString(str, 'application/xml'));
 
-    element = await fixture(html`<smv-plugin .doc=${doc}></smv-plugin>`);
+    element = await fixture(
+      html`<subscription-plugin .doc=${doc}></subscription-plugin>`
+    );
   });
 
   describe('in Publisher view', () => {
@@ -26,9 +31,9 @@ describe('Sampled Values Plugin', () => {
         await expect(element).shadowDom.to.equalSnapshot();
       });
 
-      it('the Sampled Values list looks like the latest snapshot', async () => {
+      it('the GOOSE list looks like the latest snapshot', async () => {
         await expect(
-          element.shadowRoot?.querySelector('smv-list')
+          element.shadowRoot?.querySelector('goose-list')
         ).shadowDom.to.equalSnapshot();
       });
 
@@ -37,17 +42,17 @@ describe('Sampled Values Plugin', () => {
       });
     });
 
-    describe('when selecting a Sampled Values message', () => {
-      let smv: HTMLElement;
+    describe('when selecting a GOOSE message', () => {
+      let goose: HTMLElement;
 
       beforeEach(async () => {
-        smv = Array.from(
+        goose = Array.from(
           element.shadowRoot
-            ?.querySelector('smv-list')
+            ?.querySelector('goose-list')
             ?.shadowRoot?.querySelectorAll('mwc-list-item') ?? []
-        ).filter(item => !item.noninteractive)[0];
+        ).filter(item => !item.noninteractive)[2];
 
-        (<HTMLElement>smv).click();
+        (<HTMLElement>goose).click();
         await element.updateComplete;
       });
 
@@ -59,39 +64,39 @@ describe('Sampled Values Plugin', () => {
         it('initially no ExtRefs are available in the subscriber IED', async () => {
           expect(
             element.doc.querySelectorAll(
-              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED3"], ' +
-                'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED3"]'
+              'IED[name="IED3"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED2"], ' +
+                'IED[name="IED3"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED2"]'
             ).length
           ).to.eql(0);
         });
 
         it('it looks like the latest snapshot', async () => {
-          (<HTMLElement>getItemFromSubscriberList('IED2')).click();
+          (<HTMLElement>getItemFromSubscriberList('IED3')).click();
           await element.updateComplete;
 
           await expect(getSubscriberList()).shadowDom.to.equalSnapshot();
         });
 
         it('adds the required ExtRefs to the subscriber IED', async () => {
-          (<HTMLElement>getItemFromSubscriberList('IED2')).click();
+          (<HTMLElement>getItemFromSubscriberList('IED3')).click();
 
           expect(
             element.doc.querySelectorAll(
-              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED3"], ' +
-                'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED3"]'
+              'IED[name="IED3"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED2"], ' +
+                'IED[name="IED3"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED2"]'
             ).length
-          ).to.eql(16);
+          ).to.eql(3);
         });
       });
 
-      describe('and you unsubscribe a subscribed IED', () => {
+      describe('and unsubscribing a subscribed IED', () => {
         it('initially all the ExtRefs are available in the subscriber IED', async () => {
           expect(
             element.doc.querySelectorAll(
-              'IED[name="IED1"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED3"], ' +
-                'IED[name="IED1"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED3"]'
+              'IED[name="IED1"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED2"], ' +
+                'IED[name="IED1"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED2"]'
             ).length
-          ).to.eql(16);
+          ).to.eql(3);
         });
 
         it('it looks like the latest snapshot', async () => {
@@ -106,21 +111,20 @@ describe('Sampled Values Plugin', () => {
 
           expect(
             element.doc.querySelectorAll(
-              'IED[name="IED1"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED3"], ' +
-                'IED[name="IED1"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED3"]'
+              'IED[name="IED1"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED2"], ' +
+                'IED[name="IED1"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED2"]'
             ).length
           ).to.eql(0);
         });
       });
 
-      describe('and you subscribe a partially subscribed IED', () => {
-        it('initially only 10 ExtRefs are available in the subscriber IED', async () => {
-          expect(
-            element.doc.querySelectorAll(
-              'IED[name="IED4"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED3"],' +
-                'IED[name="IED4"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED3"]'
-            ).length
-          ).to.eql(10);
+      describe('and subscribing a partially subscribed IED', () => {
+        it('initially only 2 ExtRefs are available in the subscriber IED', async () => {
+          const extRefs = element.doc.querySelectorAll(
+            'IED[name="IED4"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED2"],' +
+              'IED[name="IED4"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED2"]'
+          );
+          expect(extRefs.length).to.eql(2);
         });
 
         it('it looks like the latest snapshot', async () => {
@@ -133,12 +137,11 @@ describe('Sampled Values Plugin', () => {
         it('adds the required ExtRefs to the subscriber IED', async () => {
           (<HTMLElement>getItemFromSubscriberList('IED4')).click();
 
-          expect(
-            element.doc.querySelectorAll(
-              'IED[name="IED4"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED3"],' +
-                'IED[name="IED4"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED3"]'
-            ).length
-          ).to.eql(16);
+          const extRefs = element.doc.querySelectorAll(
+            'IED[name="IED4"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED2"],' +
+              'IED[name="IED4"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED2"]'
+          );
+          expect(extRefs.length).to.eql(5);
         });
       });
     });
@@ -147,7 +150,7 @@ describe('Sampled Values Plugin', () => {
   describe('in Subscriber view', () => {
     beforeEach(async () => {
       (<HTMLElement>(
-        element.shadowRoot?.querySelector('#smvSubscriberView')
+        element.shadowRoot?.querySelector('#gooseSubscriberView')
       )).click();
       await element.updateComplete;
     });
@@ -159,7 +162,7 @@ describe('Sampled Values Plugin', () => {
 
       it('the IED list looks like the latest snapshot', async () => {
         await expect(
-          element.shadowRoot?.querySelector('ied-list-smv')
+          element.shadowRoot?.querySelector('ied-list-goose')
         ).shadowDom.to.equalSnapshot();
       });
 
@@ -174,7 +177,7 @@ describe('Sampled Values Plugin', () => {
       beforeEach(async () => {
         ied = Array.from(
           element.shadowRoot
-            ?.querySelector('ied-list-smv')
+            ?.querySelector('ied-list-goose')
             ?.shadowRoot?.querySelectorAll('mwc-list-item') ?? []
         ).filter(item => !item.noninteractive)[1];
 
@@ -186,102 +189,100 @@ describe('Sampled Values Plugin', () => {
         await expect(getSubscriberList()).shadowDom.to.equalSnapshot();
       });
 
-      describe('and subscribing a unsubscribed Sampled Value message', () => {
+      describe('and subscribing a unsubscribed GOOSE message', () => {
         it('initially no ExtRefs are available in the subscriber IED', async () => {
           expect(
             element.doc.querySelectorAll(
-              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED3"], ' +
-                'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED3"]'
+              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED4"], ' +
+                'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED4"]'
             ).length
           ).to.eql(0);
         });
 
         it('it looks like the latest snapshot', async () => {
-          (<HTMLElement>getItemFromSubscriberList('MSVCB01')).click();
+          (<HTMLElement>getItemFromSubscriberList('IED4')).click();
           await element.updateComplete;
 
           await expect(getSubscriberList()).shadowDom.to.equalSnapshot();
         });
 
         it('adds the required ExtRefs to the subscriber IED', async () => {
-          (<HTMLElement>getItemFromSubscriberList('MSVCB01')).click();
+          (<HTMLElement>getItemFromSubscriberList('IED4')).click();
 
           expect(
             element.doc.querySelectorAll(
-              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED3"], ' +
-                'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED3"]'
+              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED4"], ' +
+                'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED4"]'
             ).length
-          ).to.eql(16);
+          ).to.eql(5);
         });
       });
 
-      describe('and unsubscribing a subscribed Sampled Value message', () => {
+      describe('and unsubscribing a subscribed GOOSE message', () => {
         beforeEach(async () => {
-          (<HTMLElement>getItemFromSubscriberList('MSVCB01')).click();
+          (<HTMLElement>getItemFromSubscriberList('IED4')).click();
           await element.updateComplete;
         });
 
         it('initially all ExtRefs are available in the subscriber IED', async () => {
           expect(
             element.doc.querySelectorAll(
-              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED3"], ' +
-                'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED3"]'
+              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED4"], ' +
+                'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED4"]'
             ).length
-          ).to.eql(16);
+          ).to.eql(5);
         });
 
         it('it looks like the latest snapshot', async () => {
-          (<HTMLElement>getItemFromSubscriberList('MSVCB01')).click();
+          (<HTMLElement>getItemFromSubscriberList('IED4')).click();
           await element.updateComplete;
 
           await expect(getSubscriberList()).shadowDom.to.equalSnapshot();
         });
 
         it('removes the required ExtRefs to the subscriber IED', async () => {
-          (<HTMLElement>getItemFromSubscriberList('MSVCB01')).click();
+          (<HTMLElement>getItemFromSubscriberList('IED4')).click();
 
-          expect(
-            element.doc.querySelectorAll(
-              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED3"], ' +
-                'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED3"]'
-            ).length
-          ).to.eql(0);
+          const extRefs = element.doc.querySelectorAll(
+            'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED4"], ' +
+              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED4"]'
+          );
+          expect(extRefs.length).to.eql(0);
         });
       });
 
-      describe('and subscribing a partially subscribed Sampled Value message', () => {
+      describe('and subscribing a partially subscribed GOOSE message', () => {
         it('initially some ExtRefs are available in the subscriber IED', async () => {
-          expect(
-            element.doc.querySelectorAll(
-              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED4"], ' +
-                'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED4"]'
-            ).length
-          ).to.eql(2);
+          const extRefs = doc.querySelectorAll(
+            'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED1"], ' +
+              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED1"]'
+          );
+          expect(extRefs.length).to.eql(4);
         });
 
         it('it looks like the latest snapshot', async () => {
-          (<HTMLElement>getItemFromSubscriberList('MSVCB02')).click();
+          (<HTMLElement>getItemFromSubscriberList('IED1')).click();
           await element.updateComplete;
 
           await expect(getSubscriberList()).shadowDom.to.equalSnapshot();
         });
 
         it('adds the required ExtRefs to the subscriber IED', async () => {
-          (<HTMLElement>getItemFromSubscriberList('MSVCB02')).click();
+          (<HTMLElement>getItemFromSubscriberList('IED1')).click();
 
           expect(
             element.doc.querySelectorAll(
-              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED4"], ' +
-                'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED4"]'
+              'IED[name="IED2"] > AccessPoint > Server > LDevice > LN0 > Inputs > ExtRef[iedName="IED1"], ' +
+                'IED[name="IED2"] > AccessPoint > Server > LDevice > LN > Inputs > ExtRef[iedName="IED1"]'
             ).length
-          ).to.eql(16);
+          ).to.eql(9);
         });
       });
     });
   });
 
   function getSubscriberList() {
-    return element.shadowRoot?.querySelector('subscriber-list-smv');
+    return element.shadowRoot?.querySelector('subscriber-list-goose');
   }
 
   function getItemFromSubscriberList(
